refactor: replace `DeleteWebUserInCustomer` with `DeleteWebUser` and update related authorizations/routes
- Removed `DeleteWebUserInCustomer` and its associated logic.
- Enhanced `DeleteWebUser` with CRM-specific authorizations and fulfillment shop handling.
- Updated routes to use `DeleteWebUser` for web user deletions.
