+++ app/Actions/Traits/Authorisations/WithCRMAuthorisation.php
@@ -39,9 +39,22 @@ trait WithCRMAuthorisation
             if (str_ends_with($routeName, '.edit') || str_ends_with($routeName, '.create')) {
                 return $this->canEdit;
             }
+
             return $request->user()->authTo(
                 [
                     "crm.{$this->shop->id}.view",
+                    "accounting.{$this->shop->organisation_id}.view"
+                ]
+            );
+        } elseif (str_starts_with($routeName, 'grp.org.fulfilments.show.crm.')) {
+            $this->canEdit = $request->user()->authTo("fulfilment-shop.{$this->shop->fulfilment->id}.edit");
+            if (str_ends_with($routeName, '.edit') || str_ends_with($routeName, '.create')) {
+                return $this->canEdit;
+            }
+
+            return $request->user()->authTo(
+                [
+                    "fulfilment-shop.{$this->shop->fulfilment->id}.view",
                     "accounting.{$this->shop->organisation_id}.view"
                 ]
             );
