+++ app/Actions/Traits/Authorisations/WithCRMEditAuthorisation.php
@@ -39,6 +39,12 @@ trait WithCRMEditAuthorisation
                     "crm.{$this->shop->id}.edit"
                 ]
             );
+        } elseif (str_starts_with($routeName, 'grp.org.fulfilments.show.crm.')) {
+            return $request->user()->authTo(
+                [
+                    "fulfilment-shop.{$this->shop->fulfilment->id}.edit",
+                ]
+            );
         }
