+++ routes/grp/web/models.php
@@ -70,8 +70,8 @@ use App\Actions\CRM\Poll\DeletePoll;
 use App\Actions\CRM\Poll\StorePoll;
 use App\Actions\CRM\Poll\UpdatePoll;
 use App\Actions\CRM\Prospect\ImportShopProspects;
+use App\Actions\CRM\WebUser\DeleteWebUser;
 use App\Actions\CRM\WebUser\StoreWebUser;
-use App\Actions\CRM\WebUser\DeleteWebUserInCustomer;
 use App\Actions\CRM\WebUser\UpdateWebUser;
 use App\Actions\Dispatching\Printer\PrintShipmentLabel;
 use App\Actions\Dispatching\Shipment\DeleteShipment;
@@ -685,7 +685,7 @@ Route::name('model_has_web_block.')->prefix('model-has-web-block')->group(functi
 });
 
 Route::patch('/web-user/{webUser:id}', UpdateWebUser::class)->name('web-user.update');
-Route::delete('/web-user/{webUser:id}', DeleteWebUserInCustomer::class)->name('web-user.delete');
+Route::delete('/web-user/{webUser:id}', DeleteWebUser::class)->name('web-user.delete');
 
 Route::name('customer.')->prefix('customer/{customer:id}')->group(function () {
     Route::post('', [StoreWebUser::class, 'inCustomer'])->name('web-user.store');
